Raw exec can use cgroups to manage PIDs

+++ client/driver/raw_exec.go
@@ -7,8 +7,11 @@ import (
 	"log"
 	"os"
 	"path/filepath"
+	"runtime"
+	"syscall"
 	"time"
 
+	multierror "github.com/hashicorp/go-multierror"
 	"github.com/hashicorp/go-plugin"
 	"github.com/hashicorp/nomad/client/allocdir"
 	"github.com/hashicorp/nomad/client/driver/env"
@@ -22,8 +25,11 @@ import (
 )
 
 const (
-	// The option that enables this driver in the Config.Options map.
-	rawExecConfigOption = "driver.raw_exec.enable"
+	// rawExecEnableOption is the option that enables this driver in the Config.Options map.
+	rawExecEnableOption = "driver.raw_exec.enable"
+
+	// rawExecNoCgroupOption forces no cgroups.
+	rawExecNoCgroupOption = "driver.raw_exec.no_cgroups"
 
 	// The key populated in Node Attributes to indicate presence of the Raw Exec
 	// driver
@@ -36,21 +42,26 @@ const (
 type RawExecDriver struct {
 	DriverContext
 	fingerprint.StaticFingerprinter
+
+	// useCgroup tracks whether we should use a cgroup to manage the process
+	// tree
+	useCgroup bool
 }
 
 // rawExecHandle is returned from Start/Open as a handle to the PID
 type rawExecHandle struct {
-	version        string
-	pluginClient   *plugin.Client
-	userPid        int
-	executor       executor.Executor
-	killTimeout    time.Duration
-	maxKillTimeout time.Duration
-	logger         *log.Logger
-	waitCh         chan *dstructs.WaitResult
-	doneCh         chan struct{}
-	taskEnv        *env.TaskEnv
-	taskDir        *allocdir.TaskDir
+	version         string
+	pluginClient    *plugin.Client
+	userPid         int
+	executor        executor.Executor
+	isolationConfig *dstructs.IsolationConfig
+	killTimeout     time.Duration
+	maxKillTimeout  time.Duration
+	logger          *log.Logger
+	waitCh          chan *dstructs.WaitResult
+	doneCh          chan struct{}
+	taskEnv         *env.TaskEnv
+	taskDir         *allocdir.TaskDir
 }
 
 // NewRawExecDriver is used to create a new raw exec driver
@@ -93,7 +104,7 @@ func (d *RawExecDriver) FSIsolation() cstructs.FSIsolation {
 
 func (d *RawExecDriver) Fingerprint(req *cstructs.FingerprintRequest, resp *cstructs.FingerprintResponse) error {
 	// Check that the user has explicitly enabled this executor.
-	enabled := req.Config.ReadBoolDefault(rawExecConfigOption, false)
+	enabled := req.Config.ReadBoolDefault(rawExecEnableOption, false)
 
 	if enabled || req.Config.DevMode {
 		d.logger.Printf("[WARN] driver.raw_exec: raw exec is enabled. Only enable if needed")
@@ -107,6 +118,14 @@ func (d *RawExecDriver) Fingerprint(req *cstructs.FingerprintRequest, resp *cstr
 }
 
 func (d *RawExecDriver) Prestart(*ExecContext, *structs.Task) (*PrestartResponse, error) {
+	// If we are on linux, running as root, cgroups are mounted, and cgroups
+	// aren't disabled by the operator use cgroups for pid management.
+	forceDisable := d.DriverContext.config.ReadBoolDefault(rawExecNoCgroupOption, false)
+	if !forceDisable && runtime.GOOS == "linux" &&
+		syscall.Geteuid() == 0 && cgroupsMounted(d.DriverContext.node) {
+		d.useCgroup = true
+	}
+
 	return nil, nil
 }
 
@@ -150,10 +169,11 @@ func (d *RawExecDriver) Start(ctx *ExecContext, task *structs.Task) (*StartRespo
 	}
 
 	execCmd := &executor.ExecCommand{
-		Cmd:            command,
-		Args:           driverConfig.Args,
-		User:           task.User,
-		TaskKillSignal: taskKillSignal,
+		Cmd:                command,
+		Args:               driverConfig.Args,
+		User:               task.User,
+		TaskKillSignal:     taskKillSignal,
+		BasicProcessCgroup: d.useCgroup,
 	}
 	ps, err := exec.LaunchCmd(execCmd)
 	if err != nil {
@@ -165,17 +185,18 @@ func (d *RawExecDriver) Start(ctx *ExecContext, task *structs.Task) (*StartRespo
 	// Return a driver handle
 	maxKill := d.DriverContext.config.MaxKillTimeout
 	h := &rawExecHandle{
-		pluginClient:   pluginClient,
-		executor:       exec,
-		userPid:        ps.Pid,
-		killTimeout:    GetKillTimeout(task.KillTimeout, maxKill),
-		maxKillTimeout: maxKill,
-		version:        d.config.Version.VersionNumber(),
-		logger:         d.logger,
-		doneCh:         make(chan struct{}),
-		waitCh:         make(chan *dstructs.WaitResult, 1),
-		taskEnv:        ctx.TaskEnv,
-		taskDir:        ctx.TaskDir,
+		pluginClient:    pluginClient,
+		executor:        exec,
+		isolationConfig: ps.IsolationConfig,
+		userPid:         ps.Pid,
+		killTimeout:     GetKillTimeout(task.KillTimeout, maxKill),
+		maxKillTimeout:  maxKill,
+		version:         d.config.Version.VersionNumber(),
+		logger:          d.logger,
+		doneCh:          make(chan struct{}),
+		waitCh:          make(chan *dstructs.WaitResult, 1),
+		taskEnv:         ctx.TaskEnv,
+		taskDir:         ctx.TaskDir,
 	}
 	go h.run()
 	return &StartResponse{Handle: h}, nil
@@ -184,11 +205,12 @@ func (d *RawExecDriver) Start(ctx *ExecContext, task *structs.Task) (*StartRespo
 func (d *RawExecDriver) Cleanup(*ExecContext, *CreatedResources) error { return nil }
 
 type rawExecId struct {
-	Version        string
-	KillTimeout    time.Duration
-	MaxKillTimeout time.Duration
-	UserPid        int
-	PluginConfig   *PluginReattachConfig
+	Version         string
+	KillTimeout     time.Duration
+	MaxKillTimeout  time.Duration
+	UserPid         int
+	PluginConfig    *PluginReattachConfig
+	IsolationConfig *dstructs.IsolationConfig
 }
 
 func (d *RawExecDriver) Open(ctx *ExecContext, handleID string) (DriverHandle, error) {
@@ -202,11 +224,19 @@ func (d *RawExecDriver) Open(ctx *ExecContext, handleID string) (DriverHandle, e
 	}
 	exec, pluginClient, err := createExecutorWithConfig(pluginConfig, d.config.LogOutput)
 	if err != nil {
+		merrs := new(multierror.Error)
+		merrs.Errors = append(merrs.Errors, err)
 		d.logger.Println("[ERR] driver.raw_exec: error connecting to plugin so destroying plugin pid and user pid")
 		if e := destroyPlugin(id.PluginConfig.Pid, id.UserPid); e != nil {
-			d.logger.Printf("[ERR] driver.raw_exec: error destroying plugin and userpid: %v", e)
+			merrs.Errors = append(merrs.Errors, fmt.Errorf("error destroying plugin and userpid: %v", e))
 		}
-		return nil, fmt.Errorf("error connecting to plugin: %v", err)
+		if id.IsolationConfig != nil {
+			ePid := pluginConfig.Reattach.Pid
+			if e := executor.ClientCleanup(id.IsolationConfig, ePid); e != nil {
+				merrs.Errors = append(merrs.Errors, fmt.Errorf("destroying resource container failed: %v", e))
+			}
+		}
+		return nil, fmt.Errorf("error connecting to plugin: %v", merrs.ErrorOrNil())
 	}
 
 	ver, _ := exec.Version()
@@ -214,17 +244,18 @@ func (d *RawExecDriver) Open(ctx *ExecContext, handleID string) (DriverHandle, e
 
 	// Return a driver handle
 	h := &rawExecHandle{
-		pluginClient:   pluginClient,
-		executor:       exec,
-		userPid:        id.UserPid,
-		logger:         d.logger,
-		killTimeout:    id.KillTimeout,
-		maxKillTimeout: id.MaxKillTimeout,
-		version:        id.Version,
-		doneCh:         make(chan struct{}),
-		waitCh:         make(chan *dstructs.WaitResult, 1),
-		taskEnv:        ctx.TaskEnv,
-		taskDir:        ctx.TaskDir,
+		pluginClient:    pluginClient,
+		executor:        exec,
+		userPid:         id.UserPid,
+		isolationConfig: id.IsolationConfig,
+		logger:          d.logger,
+		killTimeout:     id.KillTimeout,
+		maxKillTimeout:  id.MaxKillTimeout,
+		version:         id.Version,
+		doneCh:          make(chan struct{}),
+		waitCh:          make(chan *dstructs.WaitResult, 1),
+		taskEnv:         ctx.TaskEnv,
+		taskDir:         ctx.TaskDir,
 	}
 	go h.run()
 	return h, nil
@@ -232,11 +263,12 @@ func (d *RawExecDriver) Open(ctx *ExecContext, handleID string) (DriverHandle, e
 
 func (h *rawExecHandle) ID() string {
 	id := rawExecId{
-		Version:        h.version,
-		KillTimeout:    h.killTimeout,
-		MaxKillTimeout: h.maxKillTimeout,
-		PluginConfig:   NewPluginReattachConfig(h.pluginClient.ReattachConfig()),
-		UserPid:        h.userPid,
+		Version:         h.version,
+		KillTimeout:     h.killTimeout,
+		MaxKillTimeout:  h.maxKillTimeout,
+		PluginConfig:    NewPluginReattachConfig(h.pluginClient.ReattachConfig()),
+		UserPid:         h.userPid,
+		IsolationConfig: h.isolationConfig,
 	}
 
 	data, err := json.Marshal(id)
@@ -298,8 +330,15 @@ func (h *rawExecHandle) run() {
 	ps, werr := h.executor.Wait()
 	close(h.doneCh)
 	if ps.ExitCode == 0 && werr != nil {
-		if e := killProcess(h.userPid); e != nil {
-			h.logger.Printf("[ERR] driver.raw_exec: error killing user process: %v", e)
+		if h.isolationConfig != nil {
+			ePid := h.pluginClient.ReattachConfig().Pid
+			if e := executor.ClientCleanup(h.isolationConfig, ePid); e != nil {
+				h.logger.Printf("[ERR] driver.raw_exec: destroying resource container failed: %v", e)
+			}
+		} else {
+			if e := killProcess(h.userPid); e != nil {
+				h.logger.Printf("[ERR] driver.raw_exec: error killing user process: %v", e)
+			}
 		}
 	}
